requirements.txt
# Core MCP and RAG dependencies
mcp[cli]==1.0.0
langchain==0.1.0
langchain-community==0.0.20
chromadb==0.4.22

# Document processing
markitdown[all]==0.0.1

# Embeddings and ML
sentence-transformers==2.2.2
torch==2.1.2
huggingface-hub>=0.20.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
requests==2.31.0
httpx>=0.27.0

# Optional: OpenAI (if you want to use OpenAI embeddings instead of local)
# openai==1.12.0

# Optional: Flask (if you want to add a web interface later)
# Flask==3.0.0

# Optional: Tokenization (if needed for specific use cases)
# tiktoken==0.5.2 